+++ rs/pocket_ic_server/src/state_api/state.rs
@@ -280,6 +280,7 @@ pub enum PocketIcError {
     BlockmakerContainedInFailed(NodeId),
     InvalidCanisterSnapshotDirectory(String),
     CanisterSnapshotError(String),
+    InvalidTickConfigs(String),
 }
 
 impl std::fmt::Debug for OpOut {
@@ -350,6 +351,9 @@ impl std::fmt::Debug for OpOut {
             }
             OpOut::Error(PocketIcError::CanisterSnapshotError(msg)) => {
                 write!(f, "CanisterSnapshotError({msg})")
+            }
+            OpOut::Error(PocketIcError::InvalidTickConfigs(msg)) => {
+                write!(f, "InvalidTickConfigs({msg})")
             }
             OpOut::Bytes(bytes) => write!(f, "Bytes({})", base64::encode(bytes)),
             OpOut::StableMemBytes(bytes) => write!(f, "StableMemory({})", base64::encode(bytes)),