+++ rs/dogecoin/ckdoge/test_utils/src/minter.rs
@@ -2,17 +2,17 @@ use crate::events::MinterEventAssert;
 use crate::{MAX_TIME_IN_QUEUE, NNS_ROOT_PRINCIPAL};
 use candid::{Decode, Encode, Principal};
 use canlog::LogEntry;
-use ic_ckdoge_minter::candid_api::{EstimateWithdrawalFeeError, WithdrawalFee};
-use ic_ckdoge_minter::lifecycle::init::{MinterArg, UpgradeArgs};
 use ic_ckdoge_minter::{
     EstimateFeeArg, Event, EventType, Priority, Txid, UpdateBalanceArgs, UpdateBalanceError, Utxo,
     UtxoStatus,
     candid_api::{
-        GetDogeAddressArgs, RetrieveDogeOk, RetrieveDogeStatus, RetrieveDogeStatusRequest,
-        RetrieveDogeWithApprovalArgs, RetrieveDogeWithApprovalError,
+        EstimateWithdrawalFeeError, GetDogeAddressArgs, MinterInfo, RetrieveDogeOk,
+        RetrieveDogeStatus, RetrieveDogeStatusRequest, RetrieveDogeWithApprovalArgs,
+        RetrieveDogeWithApprovalError, WithdrawalFee,
     },
+    lifecycle::init::{MinterArg, UpgradeArgs},
 };
-use ic_management_canister_types::CanisterId;
+use ic_management_canister_types::{CanisterId, CanisterStatusResult};
 use ic_metrics_assert::{MetricsAssert, PocketIcHttpQuery};
 use icrc_ledger_types::icrc1::account::Account;
 use pocket_ic::common::rest::RawMessageId;
@@ -112,6 +112,32 @@ impl MinterCanister {
     ) -> Result<std::vec::Vec<u8>, RejectResponse> {
         self.env
             .update_call(self.id, sender, "update_balance", Encode!(args).unwrap())
+    }
+
+    pub fn get_canister_status(&self) -> CanisterStatusResult {
+        let call_result = self
+            .env
+            .update_call(
+                self.id,
+                Principal::anonymous(),
+                "get_canister_status",
+                Encode!().unwrap(),
+            )
+            .expect("BUG: failed to call get_canister_status");
+        Decode!(&call_result, CanisterStatusResult).unwrap()
+    }
+
+    pub fn get_minter_info(&self) -> MinterInfo {
+        let call_result = self
+            .env
+            .update_call(
+                self.id,
+                Principal::anonymous(),
+                "get_minter_info",
+                Encode!().unwrap(),
+            )
+            .expect("BUG: failed to call get_minter_info");
+        Decode!(&call_result, MinterInfo).unwrap()
     }
 
     pub fn estimate_withdrawal_fee(