+++ rs/pocket_ic_server/src/main.rs
@@ -65,7 +65,7 @@ static MAINNET_ROUTING_TABLE: &[u8] = include_bytes!("mainnet_routing_table.json
 
 #[derive(Parser)]
 #[clap(name = "pocket-ic-server")]
-#[clap(version = "10.0.0")]
+#[clap(version = "11.0.0")]
 struct Args {
     /// The IP address to which the PocketIC server should bind (defaults to 127.0.0.1)
     #[clap(long, short)]