+++ packages/pocket-ic/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## 11.0.0 - 2025-12-05
+
 ### Added
 - The function `PocketIcBuilder::with_initial_time` to specify the initial timestamp of the newly created PocketIC instance.
 - The parameter `ttl` to `StartServerParams` to specify the TTL of the PocketIC server.